requirements.txt
# Florida Retirement Resources Multi-Agent System
# Python dependencies

# Google ADK (Agent Development Kit)
google-adk>=0.1.0

# Google Generative AI
google-genai>=0.2.0

# Core dependencies (usually included with ADK, but listed for clarity)
requests>=2.31.0

# Optional: For enhanced functionality
# python-dotenv>=1.0.0  # For environment variable management
# pydantic>=2.0.0  # For data validation (usually included with ADK)

